+++ examples/.experimental/next-app-dir/src/app/posts/_lib/trpc.ts
@@ -17,3 +17,8 @@ export const nextProc = t.procedure
       normalizeFormData: true,
     }),
   );
+
+export {
+  experimental_notFound as notFound,
+  experimental_redirect as redirect,
+} from '@trpc/server/adapters/next-app-dir';
